requirements.txt
django
markdown
djangorestframework
markdown
django-filter
pillow
pygments